fix all warnings

+++ include/rv32i/rvi_rv32i_type_i_system.hpp
@@ -74,7 +74,7 @@ private:
     ExecutionStatus Exit(InterpreterState* state) const {
         std::fflush(stdout);
         state->pc += 4u;
-        state->return_code = state->regs.Get(10); // a0
+        state->return_code = static_cast<int32_t>(state->regs.Get(10)); // a0
         return ExecutionStatus::Exit;
     }
